sendEther uint64->uint

+++ GatewayVote.sol
@@ -64,7 +64,7 @@ contract GatewayVote
     event GatewayAddrChanged(uint32 appCode, address newer, uint256 indexed operation);
 
     event RecvEther(address sender,string targetChain,string targetAddr);
-    event SendEther(address sender,address receiver,uint64 amount,string proposal);
+    event SendEther(address sender,address receiver,uint amount,string proposal);
 
     // METHODS
 
@@ -339,7 +339,7 @@ contract GatewayVote
         emit RecvEther(msg.sender,targetChain,targetAddr);
         return true;
     }
-    function sendEther(address receiver,uint64 amount,string proposal) external returns (bool){
+    function sendEther(address receiver,uint amount,string proposal) external returns (bool){
         require(isVoter(tx.origin) && !mStopped);
         require(amount <= address(this).balance);
